修改代码生成器实现,简化Mapper/Service

+++ src/main/java/com/lmt/code/bean/ColumnBean.java
@@ -24,7 +24,7 @@ public class ColumnBean implements Serializable {
     /**
      * 数据长度
      */
-    private int dataLength;
+    private long dataLength;
 
     /**
      * 字段名
